cephadm: move nvmeof class to a new file
