+++ src/cephadm/cephadmlib/daemons/__init__.py
@@ -1,5 +1,12 @@
 from .custom import CustomContainer
 from .tracing import Tracing
 from .ingress import HAproxy, Keepalived
+from .nvmeof import CephNvmeof
 
-__all__ = ['CustomContainer', 'Tracing', 'HAproxy', 'Keepalived']
+__all__ = [
+    'CephNvmeof',
+    'CustomContainer',
+    'HAproxy',
+    'Keepalived',
+    'Tracing',
+]
